feat!: do not require calling setup
Setup the plugin in the plugin directory which automatically executes.
There is no need to call setup if the user is perfectly happy with the
default options
